pyproject.toml
[project]
name = "mx-templates-bulk-addressing"
version = "0.1.0"
description = "Python tool for automating Cisco Meraki MX VLAN configurations using Excel for IP addressing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.6.3",
    "cryptography>=45.0.6",
    "meraki>=2.0.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
